+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## Version 1.2.0
+## Version 1.2.1
 - New message requests system: Messages from users you don't follow appear in a separate requests tab
 - Enhanced breathing exercises with a ready button
 - Connect 4 game improvements
@@ -15,6 +15,12 @@
 - Privacy chnages
 - Privacy Policy, CookiesPolicy, Community Guidlines and Safety, Account management, Terms of Service and About Updated.
 - Policies Checkbox added on the Register page.
+- New Device Registration Update
+- Save Changes bug fixed in the Side Room Components
+- Added a FAQ Page
+- Room Password Updated
+- Deactivation/Deletion updates, Users who delete their account all their activities and details are deleted on firestore
+- Users who deactivate their account, their activities and profile still in tact but will be logged out until they reactivate.
 
 ## Version 1.0.0
 - Initial release of SideEye